+++ spec/tokenizer_spec.rb
@@ -171,6 +171,12 @@ describe Dentaku::Tokenizer do
     expect(tokens.map(&:value)).to eq(['size', :lt, 3, :or, 'admin', :eq, 1])
   end
 
+  it 'tokenizes square brackets for data structure access' do
+    tokens = tokenizer.tokenize('a[1]')
+    expect(tokens.map(&:category)).to eq(%i(identifier access numeric access))
+    expect(tokens.map(&:value)).to eq(['a', :lbracket, 1, :rbracket])
+  end
+
   it 'detects unbalanced parentheses' do
     expect { tokenizer.tokenize('(5+3') }.to raise_error(Dentaku::TokenizerError, /too many opening parentheses/)
     expect { tokenizer.tokenize(')')    }.to raise_error(Dentaku::TokenizerError, /too many closing parentheses/)